Refactor settings screens: extract hooks, components, and centralize constants
- Extract ConnectionSettings and SyncStrategySettings from SettingsScreen (983 → 239 lines)
- Create useAutoSave hook replacing duplicated debounce/save/restart pattern across 3 screens
- Create useTimeout hook replacing scattered timeout ref + cleanup patterns
- Centralize constants (timing, thresholds, URLs) into src/constants/index.ts
- Fix SettingsService silent failures: throw on error instead of returning false
- Add export format registry eliminating switch statements in ExportDataScreen
