+++ src/components/index.ts
@@ -46,3 +46,5 @@ export { DatabaseStatistics } from "./features/databases/DatabaseStatistics"
 export { ServerConnection } from "./features/settings/ServerConnection"
 export { StatsCard } from "./features/settings/StatsCard"
 export { PresetOption } from "./features/settings/PresetOption"
+export { ConnectionSettings } from "./features/settings/ConnectionSettings"
+export { SyncStrategySettings } from "./features/settings/SyncStrategySettings"
