+++ src/constants/index.ts
@@ -1,0 +1,20 @@
+/**
+ * Copyright (C) 2026 Max Dietrich
+ * Licensed under the GNU AGPLv3. See LICENSE in the project root for details.
+ */
+
+// Timing
+export const AUTOSAVE_DEBOUNCE_MS = 1500
+export const STATS_REFRESH_IDLE = 30_000
+export const STATS_REFRESH_FAST = 3_000
+export const SERVER_TIMEOUT = 5_000
+export const SERVER_CHECK_INTERVAL = 5 * 60 * 1000
+export const SAVE_SUCCESS_DISPLAY_MS = 2000
+
+// Thresholds
+export const HIGH_QUEUE_THRESHOLD = 50
+export const CRITICAL_QUEUE_THRESHOLD = 100
+
+// URLs
+export const REPO_URL = "https://github.com/dietrichmax/colota"
+export const PRIVACY_POLICY_URL = "https://mxd.codes/colota/privacy-policy"
