+++ src/services/SettingsService.ts
@@ -12,72 +12,73 @@ import { Settings } from "../types/global"
  */
 export const SettingsService = {
   /**
-   * Saves a single setting to native storage
-   * Handles type conversion and key mapping
+   * Saves a single setting to native storage.
+   * Handles type conversion and key mapping.
+   * Throws on failure so callers can handle errors explicitly.
    */
-  updateSetting: async (key: keyof Settings, value: any): Promise<boolean> => {
-    try {
-      let stringValue: string
+  updateSetting: async (key: keyof Settings, value: any): Promise<void> => {
+    let stringValue: string
 
-      switch (key) {
-        case "interval":
-          // UI: seconds → Native: milliseconds
-          stringValue = (Number(value) * 1000).toString()
-          break
+    switch (key) {
+      case "interval":
+        // UI: seconds → Native: milliseconds
+        stringValue = (Number(value) * 1000).toString()
+        break
 
-        case "fieldMap":
-        case "customFields":
-          // Object → JSON string
-          stringValue = JSON.stringify(value)
-          break
+      case "fieldMap":
+      case "customFields":
+        // Object → JSON string
+        stringValue = JSON.stringify(value)
+        break
 
-        case "distance":
-          // JS key: 'distance' → Native key: 'minUpdateDistance'
-          await NativeLocationService.saveSetting("minUpdateDistance", value.toString())
-          // Also save as 'distance' for JS hydration
-          stringValue = value.toString()
-          break
+      case "distance":
+        // JS key: 'distance' → Native key: 'minUpdateDistance'
+        await NativeLocationService.saveSetting("minUpdateDistance", value.toString())
+        // Also save as 'distance' for JS hydration
+        stringValue = value.toString()
+        break
 
-        case "syncInterval":
-        case "retryInterval":
-        case "maxRetries":
-        case "accuracyThreshold":
-          // Already in correct format (numbers)
-          stringValue = String(value)
-          break
+      case "syncInterval":
+      case "retryInterval":
+      case "maxRetries":
+      case "accuracyThreshold":
+        // Already in correct format (numbers)
+        stringValue = String(value)
+        break
 
-        case "filterInaccurateLocations":
-        case "isOfflineMode":
-          // Boolean → "true"/"false"
-          stringValue = String(value)
-          break
+      case "filterInaccurateLocations":
+      case "isOfflineMode":
+        // Boolean → "true"/"false"
+        stringValue = String(value)
+        break
 
-        default:
-          // Handles 'endpoint', 'syncPreset', etc.
-          stringValue = String(value)
-      }
-
-      await NativeLocationService.saveSetting(key, stringValue)
-      return true
-    } catch (err) {
-      console.error(`[SettingsService] Failed to save ${key}:`, err)
-      return false
+      default:
+        // Handles 'endpoint', 'syncPreset', etc.
+        stringValue = String(value)
     }
+
+    await NativeLocationService.saveSetting(key, stringValue)
   },
 
   /**
-   * Updates multiple settings in parallel
+   * Updates multiple settings in parallel.
+   * Throws if any individual setting fails to persist.
    */
   updateMultiple: async (settingsUpdate: Partial<Settings>): Promise<void> => {
-    const keys = Object.keys(settingsUpdate)
-    console.log(`[SettingsService] Batch updating ${keys.length} settings`)
+    const entries = Object.entries(settingsUpdate)
+    console.log(`[SettingsService] Batch updating ${entries.length} settings`)
 
-    const promises = Object.entries(settingsUpdate).map(([key, val]) =>
-      SettingsService.updateSetting(key as keyof Settings, val)
+    const results = await Promise.allSettled(
+      entries.map(([key, val]) => SettingsService.updateSetting(key as keyof Settings, val))
     )
 
-    await Promise.all(promises)
-    console.log("[SettingsService] ✅ Batch update completed")
+    const failures = results.filter((r) => r.status === "rejected")
+    if (failures.length > 0) {
+      failures.forEach((f) => console.error("[SettingsService] Setting failed:", (f as PromiseRejectedResult).reason))
+      throw new Error(`Failed to save ${failures.length} of ${entries.length} settings`)
+    }
+
+    console.log("[SettingsService] Batch update completed")
   }
 }
